Switched to using outside Module to clean up code.
